Removed ItemProcessorManagerInterface
