+++ layers/Engine/packages/component-model/src/CheckpointProcessors/CheckpointProcessorManagerInterface.php
@@ -4,8 +4,11 @@ declare(strict_types=1);
 
 namespace PoP\ComponentModel\CheckpointProcessors;
 
-use PoP\ComponentModel\ItemProcessors\ItemProcessorManagerInterface;
-
-interface CheckpointProcessorManagerInterface extends ItemProcessorManagerInterface
+interface CheckpointProcessorManagerInterface
 {
+    public function getLoadedItemFullNameProcessorInstances();
+    public function getLoadedItems();
+    public function overrideProcessorClass(string $overrideClass, string $withClass, array $forItemNames): void;
+    public function getItemProcessor(array $item);
+    public function getProcessor(array $item);
 }
