+++ layers/Engine/packages/component-model/src/ModuleProcessors/ModuleProcessorManagerInterface.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace PoP\ComponentModel\ModuleProcessors;
 
-use PoP\ComponentModel\ItemProcessors\ItemProcessorManagerInterface;
-
-interface ModuleProcessorManagerInterface extends ItemProcessorManagerInterface
+interface ModuleProcessorManagerInterface
 {
-
+    public function getLoadedItemFullNameProcessorInstances();
+    public function getLoadedItems();
+    public function overrideProcessorClass(string $overrideClass, string $withClass, array $forItemNames): void;
+    public function getItemProcessor(array $item);
+    public function getProcessor(array $item);
 }
